+++ CHANGELOG.md
@@ -26,6 +26,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - Fix slider drag and drop issues
 - Handle 422 error for seats API
 - Fix storybook and webpack dependencies
+- Fix crowdin duplication id for batch order seats
 
 ## [3.4.0] - 2026-03-17